created webpack.config.dev.js and set webpack dev server script to require an env
